+++ src/indicators/efficiency_ratio.rs
@@ -3,6 +3,8 @@ use std::fmt;
 
 use crate::errors::*;
 use crate::traits::{Close, Next, Reset};
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
 
 /// Kaufman's Efficiency Ratio (ER).
 ///
@@ -28,6 +30,7 @@ use crate::traits::{Close, Next, Reset};
 /// assert_eq!(er.next(19.0), 0.75);
 /// ```
 
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub struct EfficiencyRatio {
     length: u32,
     prices: VecDeque<f64>,
@@ -83,10 +86,10 @@ impl Next<f64> for EfficiencyRatio {
     }
 }
 
-impl<'a, T: Close> Next<&'a T> for EfficiencyRatio {
+impl<T: Close> Next<&T> for EfficiencyRatio {
     type Output = f64;
 
-    fn next(&mut self, input: &'a T) -> f64 {
+    fn next(&mut self, input: &T) -> f64 {
         self.next(input.close())
     }
 }