+++ src/indicators/rate_of_change.rs
@@ -3,6 +3,8 @@ use core::fmt;
 
 use crate::errors::*;
 use crate::traits::{Close, Next, Reset};
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
 
 /// Rate of Change (ROC)
 ///
@@ -37,6 +39,7 @@ use crate::traits::{Close, Next, Reset};
 ///
 /// * [Rate of Change, Wikipedia](https://en.wikipedia.org/wiki/Momentum_(technical_analysis))
 ///
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 #[derive(Debug, Clone)]
 pub struct RateOfChange {
     length: u32,
@@ -80,10 +83,10 @@ impl Next<f64> for RateOfChange {
     }
 }
 
-impl<'a, T: Close> Next<&'a T> for RateOfChange {
+impl<T: Close> Next<&T> for RateOfChange {
     type Output = f64;
 
-    fn next(&mut self, input: &'a T) -> f64 {
+    fn next(&mut self, input: &T) -> f64 {
         self.next(input.close())
     }
 }